% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_configurations.R
\name{add_description}
\alias{add_description}
\title{Add a description}
\usage{
add_description(dataset_uid, field, description)
}
\arguments{
\item{dataset_uid}{dataset_uid}

\item{field}{field_id}

\item{description}{description as string}
}
\value{
processor_uid
}
\description{
Add a description
}
