% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_file_to_dataset.R
\name{add_file_to_dataset}
\alias{add_file_to_dataset}
\title{Add CSV File to dataset}
\usage{
add_file_to_dataset(filepath, dataset_uid, encoding)
}
\arguments{
\item{filepath}{path to csv file}

\item{dataset_uid}{dataset_uid}

\item{encoding}{encdoing of the csv file}
}
\value{
resource_uid
}
\description{
Adds a local CSV file to a dataset on ODS (only possible with automation API)
}
