% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_metadata.R
\name{add_metadata}
\alias{add_metadata}
\title{Add Metadata field}
\usage{
add_metadata(
  dataset_uid,
  template = NULL,
  meta_name = NULL,
  meta_value = NULL,
  meta_df = NULL
)
}
\arguments{
\item{dataset_uid}{the internal uid of the dataset. The dataset_uid can be taken from the catalog}

\item{template}{kann metadata_catalog entnommen werden}

\item{meta_name}{kann metadata_catalog entnommen werden}

\item{meta_value}{neuer Wert}

\item{meta_df}{4}
}
\value{
TRUE or NULL
}
\description{
Adds a metadata field (might be deprecated in the future)
}
