% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_metadata_from_scheme.R
\name{add_metadata_from_scheme}
\alias{add_metadata_from_scheme}
\title{add_metadata_from_scheme}
\usage{
add_metadata_from_scheme(filepath, save_local = TRUE)
}
\arguments{
\item{filepath}{Pfad zum ausgefüllten Schema}

\item{save_local}{siehe get_dataset_info}
}
\value{
dataset_uid
}
\description{
Erstellt neuen Datensatz mit entsprechender Kennung und befüllt diesen mit Metadaten
Funktion kann nur auf genau definiertes Schema angewendet werden.
Alle Metadaten im Schema werden auf das Portal gesladen.
}
