% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_timeserie_precision.R
\name{add_timeserie_precision}
\alias{add_timeserie_precision}
\title{Add timeserie precision}
\usage{
add_timeserie_precision(dataset_uid, field, timeserie)
}
\arguments{
\item{dataset_uid}{the internal uid of the dataset. The dataset_uid can be taken from the catalog}

\item{field}{ID of field (variablename in the data.frame)}

\item{timeserie}{timeserie (check opendatasoft for available timeseries)}
}
\value{
processor_uid
}
\description{
Add timeserie precision
}
