% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_type.R
\name{add_type}
\alias{add_type}
\title{Change the data type of a field}
\usage{
add_type(dataset_uid, field, type)
}
\arguments{
\item{dataset_uid}{the internal uid of the dataset. The dataset_uid can be taken from the catalog}

\item{field}{ID of field (variablename in the data.frame)}

\item{type}{type (check opendatasoft for available types)}
}
\value{
processor_uid
}
\description{
Change the data type of a field
}
