% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_configurations.R
\name{add_unit}
\alias{add_unit}
\title{Add a Unit}
\usage{
add_unit(dataset_uid, field, unit)
}
\arguments{
\item{dataset_uid}{dataset_uid}

\item{field}{field_id}

\item{unit}{unit (check opendatasoft for available units)}
}
\value{
processor_uid
}
\description{
Add a Unit
}
