% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_identifier.R
\name{change_identifier}
\alias{change_identifier}
\title{Change the field ID}
\usage{
change_identifier(dataset_uid, field, to_name)
}
\arguments{
\item{dataset_uid}{the internal uid of the dataset. The dataset_uid can be taken from the catalog}

\item{field}{ID of field (variablename in the data.frame)}

\item{to_name}{new_field_id}
}
\value{
processor_uid
}
\description{
Change the field ID
}
