% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_visibility.R
\name{change_visibility}
\alias{change_visibility}
\title{Change visibility of a dataset}
\usage{
change_visibility(dataset_uid, restrict)
}
\arguments{
\item{dataset_uid}{the internal uid of the dataset. The dataset_uid can be taken from the catalog}

\item{restrict}{if `TRUE`, the dataset will not be visible for everyone anymore. If `FALSE`, restrictions on visibility are lifted.}
}
\description{
Restricts the visibility of a dataset. If restrict is set to `TRUE`, only logged in user with appropriate permissions can see the dataset.
}
