% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dataset_exist.R
\name{check_dataset_exist}
\alias{check_dataset_exist}
\title{check_dataset_exist}
\usage{
check_dataset_exist(dataset_uid)
}
\arguments{
\item{dataset_uid}{the internal uid of the dataset. The dataset_uid can be taken from the catalog}
}
\value{
boolean: wenn Datensatz existiert TRUE sonst FALSE
}
\description{
Überprüft ob ein Datensatz mit der gegebenen dataset_uid existiert
}
