% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_dataset.R
\name{copy_dataset}
\alias{copy_dataset}
\title{Copy a dataset}
\usage{
copy_dataset(dataset_uid)
}
\arguments{
\item{dataset_uid}{dataset_uid}
}
\value{
metadata of the newly created dataset
}
\description{
Copy all data and metadata from one dataset to another
}
