% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataset_from_df.R
\name{create_dataset_from_df}
\alias{create_dataset_from_df}
\title{Create new dataset from metadata data.frame}
\usage{
create_dataset_from_df(metadata_df, id_part)
}
\arguments{
\item{metadata_df}{a data.frame as received from get_metadata()}

\item{id_part}{the dataset id without the number (for example "sk-stat")}
}
\value{
dataset_uid from the created dataset
}
\description{
With this function you can easily copy and customize metadata from existing sources to create a new dataset. You can call the get_metadata() function, edit the value you want and set the new metadata directly from the metadata_df.
}
