% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataset_from_list.R
\name{create_dataset_from_list}
\alias{create_dataset_from_list}
\title{Create new dataset from metadata list}
\usage{
create_dataset_from_list(metadata_list, id_part)
}
\arguments{
\item{id_part}{the dataset id without the number (for example "sk-stat")}

\item{metadatalist}{a list as received from get_metadata()}
}
\value{
dataset_uid from the created dataset
}
\description{
With this function you can easily copy and customize metadata from existing sources to create a new dataset. You can call the get_metadata() function, edit the value you want and set the new metadata directly from the metadata_df.
}
