% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_new_dataset_id.R
\name{create_new_dataset_id}
\alias{create_new_dataset_id}
\title{create_new_dataset_id}
\usage{
create_new_dataset_id(test_id, save_local = TRUE, metadata_cat = NULL)
}
\arguments{
\item{test_id}{String in Form Departement-Amt (Beispiel: sk-stat)}

\item{save_local}{if `TRUE`, the catalog will be saved locally}

\item{metadata_cat}{Metadaten Katalog (default ist `NULL`)}

\item{dataset_uid}{the internal uid of the dataset. The dataset_uid can be taken from the catalog}

\item{schema}{path to the metadata xlsx schema}

\item{filepath}{path to the csv file where the actual data is stored}

\item{resource_uid}{the internal uid of a specific resource (namely a csv file or an attachment)}

\item{change_uid}{the internal uid of a change}

\item{encoding}{File encoding (for example utf-8)

Attachments}

\item{directory}{Pfad zu Ordner welcher alle Anlagen enthält (Wenn nicht NULL, werden alle Inhalte des Ordners hochgeladen)}

\item{files}{eines oder mehrere Files die als Anlage hochgeladen werden

Metadata}

\item{template}{kann metadata_catalog entnommen werden}

\item{meta_name}{kann metadata_catalog entnommen werden}

\item{meta_value}{neuer Wert}
}
\value{
dataset_id/technische Kennung
}
\description{
Funktion erzeugt Kennung mit durchgängiger Nummerierung ohne Duplikate
}
