% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ogd.R
\name{create_ogd}
\alias{create_ogd}
\title{Create a ODS Dataset}
\usage{
create_ogd(data, schema, filepath, attachments = NULL)
}
\arguments{
\item{data}{R Dataframe}

\item{schema}{path to the metadata xlsx schema}

\item{filepath}{path to the csv file where the actual data is stored}

\item{attachments}{one or several files or a directory (see `add_attachments`)}
}
\description{
Takes a data.frame and a metadata schema and creates a complete dataset on ODS
}
