% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_attachments.R
\name{delete_attachments}
\alias{delete_attachments}
\title{delete_attachments}
\usage{
delete_attachments(dataset_uid)
}
\arguments{
\item{dataset_uid}{the internal uid of the dataset. The dataset_uid can be taken from the catalog}
}
\description{
Funktion um Datensatz Anhänge zu löschen
}
