% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_dataset.R
\name{delete_dataset}
\alias{delete_dataset}
\title{Datensatz löschen}
\usage{
delete_dataset(dataset_uid, ask = TRUE)
}
\arguments{
\item{dataset_uid}{the internal uid of the dataset. The dataset_uid can be taken from the catalog}

\item{ask}{wenn TRUE wird nach Bestätigung durch USer gefragt.}

\item{schema}{path to the metadata xlsx schema}

\item{filepath}{path to the csv file where the actual data is stored}

\item{save_local}{if `TRUE`, the catalog will be saved locally}

\item{resource_uid}{the internal uid of a specific resource (namely a csv file or an attachment)}

\item{change_uid}{the internal uid of a change}

\item{encoding}{File encoding (for example utf-8)

Attachments}

\item{directory}{Pfad zu Ordner welcher alle Anlagen enthält (Wenn nicht NULL, werden alle Inhalte des Ordners hochgeladen)}

\item{files}{eines oder mehrere Files die als Anlage hochgeladen werden

Metadata}

\item{template}{kann metadata_catalog entnommen werden}

\item{meta_name}{kann metadata_catalog entnommen werden}

\item{meta_value}{neuer Wert}
}
\description{
Funktion um Datensatz zu löschen
}
