% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_resource.R
\name{delete_resource}
\alias{delete_resource}
\title{delete_resource}
\usage{
delete_resource(dataset_uid, resource_uid = NULL)
}
\arguments{
\item{dataset_uid}{the internal uid of the dataset. The dataset_uid can be taken from the catalog}

\item{resource_uid}{the internal uid of a specific resource (namely a csv file or an attachment)}
}
\description{
Resource eines Datensatzes löschen
}
