% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_resource.R
\name{download_resource}
\alias{download_resource}
\title{Download dataset resource}
\usage{
download_resource(dataset_uid, resource_uid = NULL, encoding = "UTF-8")
}
\arguments{
\item{dataset_uid}{the internal uid of the dataset. The dataset_uid can be taken from the catalog}

\item{resource_uid}{the internal uid of a specific resource (namely a csv file or an attachment)}

\item{encoding}{File encoding (for example utf-8)}
}
\value{
Downloaded (CSV-) resource
}
\description{
Download dataset resource
}
