% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset_attachments.R
\name{get_dataset_attachments}
\alias{get_dataset_attachments}
\title{get_dataset_attachments}
\usage{
get_dataset_attachments(dataset_uid)
}
\arguments{
\item{dataset_uid}{the internal uid of the dataset. The dataset_uid can be taken from the catalog}
}
\value{
Datensatz mit allen Datenquellen die dem entsprechenden ODS Datensatz zugeordnet sind
}
\description{
Alle Anhänge eines Datensatzes anzeigen
}
