% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset_info.R
\name{get_dataset_info}
\alias{get_dataset_info}
\title{get_dataset_info}
\usage{
get_dataset_info(save_local = TRUE, dataset_uid = NULL)
}
\arguments{
\item{save_local}{logical ob Katalog lokal gespeichert werden soll (default ist TRUE)}

\item{dataset_uid}{dataset_uid wenn dataset_info von einem bestimmten dataset heruntergeladen werden soll}
}
\description{
Funktion um alle derzeit aufgeschalteten Metadaten herunterzuladen. Hier kann auch die dataset_uid entnommen werden.
Der Catalog wird automatisch in der Variable metadata_catalog gespeichert
}
