% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset_resource.R
\name{get_dataset_resource}
\alias{get_dataset_resource}
\title{get_dataset_resource}
\usage{
get_dataset_resource(dataset_uid)
}
\arguments{
\item{dataset_uid}{the internal uid of the dataset. The dataset_uid can be taken from the catalog}
}
\value{
Datensatz mit allen Datenquellen die dem entsprechenden ODS Datensatz zugeordnet sind
}
\description{
get_dataset_resource
}
