% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset_status.R
\name{get_dataset_status}
\alias{get_dataset_status}
\title{get_dataset_status}
\usage{
get_dataset_status(dataset_uid)
}
\arguments{
\item{dataset_uid}{the internal uid of the dataset. The dataset_uid can be taken from the catalog}
}
\value{
Name des dataset_status
}
\description{
Datensatz Status herunterladen
}
