% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fields.R
\name{get_fields}
\alias{get_fields}
\title{Get dataset fields}
\usage{
get_fields(dataset_uid)
}
\arguments{
\item{dataset_uid}{the internal uid of the dataset. The dataset_uid can be taken from the catalog}
}
\value{
a data.frame containing the fields informatin
}
\description{
Retrieves the fields of the dataset including name, type, annotations and description
}
