% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latest_change.R
\name{get_latest_change}
\alias{get_latest_change}
\title{get_latest_change}
\usage{
get_latest_change(dataset_uid)
}
\arguments{
\item{dataset_uid}{the internal uid of the dataset. The dataset_uid can be taken from the catalog}
}
\value{
change_uid
}
\description{
Neueste change_uid
}
