% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_number_of_datasets.R
\name{get_number_of_datasets}
\alias{get_number_of_datasets}
\title{Get Number of Datasets}
\usage{
get_number_of_datasets()
}
\value{
number of datasets on ODS
}
\description{
Retrieves the number of all datasets on ODS (including unpublished ones)
Works only correctly with Automation API.
}
