% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_field_sortable.R
\name{make_field_sortable}
\alias{make_field_sortable}
\title{Make field sortable}
\usage{
make_field_sortable(dataset_uid, field)
}
\arguments{
\item{dataset_uid}{the internal uid of the dataset. The dataset_uid can be taken from the catalog}

\item{field}{ID of field (variablename in the data.frame)}
}
\value{
processor_uid
}
\description{
Make field sortable
}
