% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_configurations.R
\name{mark_unique_id}
\alias{mark_unique_id}
\title{Mark field as unique ID}
\usage{
mark_unique_id(dataset_uid, field)
}
\arguments{
\item{dataset_uid}{dataset_uid}

\item{field}{field_id}
}
\value{
processor_uid
}
\description{
Mark field as unique ID
}
