% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odsAutomationR-package.R
\docType{package}
\name{odsAutomationR-package}
\alias{odsAutomationR}
\alias{odsAutomationR-package}
\title{odsAutomationR: ODS Automation API Wrapper}
\description{
Accessing the Automation API of Opendatasoft. Uploading data and metadata, publishing and deleting datasets, etc.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://ogd.tg.ch}
  \item Report bugs at \url{http://ogd.tg.ch}
}

}
\author{
\strong{Maintainer}: Felix Lorenz \email{ogd@tg.ch}

}
\keyword{internal}
