% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_configurations.R
\name{order_fields}
\alias{order_fields}
\title{Reorder fields}
\usage{
order_fields(dataset_uid, order_vec)
}
\arguments{
\item{dataset_uid}{dataset_uid}

\item{order_vec}{a vevtor of field_ids in the new order}
}
\value{
processor_uid
}
\description{
Reorder fields
}
