% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_response.R
\name{prepare_response}
\alias{prepare_response}
\title{Prepare API Response}
\usage{
prepare_response(res)
}
\arguments{
\item{res}{API response}
}
\value{
data.frame or  list from json
}
\description{
Prepares JSON and converts Umlaute correctly
}
