% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_configurations.R
\name{rename_field}
\alias{rename_field}
\title{Rename field}
\usage{
rename_field(dataset_uid, field, to_name)
}
\arguments{
\item{dataset_uid}{dataset_uid}

\item{field}{field_id}

\item{to_name}{new name}
}
\value{
processor_uid
}
\description{
Rename field
}
