% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore_change.R
\name{restore_change}
\alias{restore_change}
\title{restore_change}
\usage{
restore_change(dataset_uid, change_uid)
}
\arguments{
\item{dataset_uid}{the internal uid of the dataset. The dataset_uid can be taken from the catalog}

\item{change_uid}{the internal uid of a change}
}
\description{
Historie wiederherstellen
}
