% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reupload_resource.R
\name{reupload_resource}
\alias{reupload_resource}
\title{Delete and reupload resource}
\usage{
reupload_resource(dataset_uid, resource_uid = NULL, filepath, encoding)
}
\arguments{
\item{dataset_uid}{the internal uid of the dataset. The dataset_uid can be taken from the catalog}

\item{resource_uid}{the internal uid of a specific resource (namely a csv file or an attachment)}

\item{filepath}{path to the csv file where the actual data is stored}

\item{encoding}{File encoding (for example utf-8)}
}
\description{
Delete and reupload resource
}
