% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_metadata.R
\name{set_metadata}
\alias{set_metadata}
\title{set_metadata}
\usage{
set_metadata(dataset_uid, template, meta_name, meta_value)
}
\arguments{
\item{dataset_uid}{the internal uid of the dataset. The dataset_uid can be taken from the catalog}

\item{template}{kann metadata_catalog entnommen werden}

\item{meta_name}{kann metadata_catalog entnommen werden}

\item{meta_value}{neuer Wert}

\item{schema}{path to the metadata xlsx schema}

\item{filepath}{path to the csv file where the actual data is stored}

\item{save_local}{if `TRUE`, the catalog will be saved locally}

\item{resource_uid}{the internal uid of a specific resource (namely a csv file or an attachment)}

\item{change_uid}{the internal uid of a change}

\item{encoding}{File encoding (for example utf-8)

Attachments}

\item{directory}{Pfad zu Ordner welcher alle Anlagen enthält (Wenn nicht NULL, werden alle Inhalte des Ordners hochgeladen)}

\item{files}{eines oder mehrere Files die als Anlage hochgeladen werden

Metadata}
}
\value{
`TRUE` if call was successfull, `NULL` otherwise
}
\description{
Funktion um einen Eintrag in den Metadaten zu ändern
}
