% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_attachments.R
\name{update_attachments}
\alias{update_attachments}
\title{update_attachments}
\usage{
update_attachments(directory = NULL, files = NULL, dataset_uid)
}
\arguments{
\item{encoding}{File encoding}
}
\description{
Funktion um Datensatz Anhänge zu löschen und neue hochzuladen
}
