% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_dataset_from_df.R
\name{update_dataset_from_df}
\alias{update_dataset_from_df}
\title{Update dataset with metadata data.frame}
\usage{
update_dataset_from_df(
  dataset_uid,
  metadata_df,
  id_part = NULL,
  dataset_id = NULL,
  is_restricted = TRUE
)
}
\arguments{
\item{dataset_uid}{the internal uid of the dataset. The dataset_uid can be taken from the catalog}

\item{metadata_df}{a data.frame as received from get_metadata()}

\item{id_part}{the dataset id without the number (for example "sk-stat")}

\item{dataset_id}{full dataset_id (e.g. sk-stat-100). It is recommendend to just provide the id_part}

\item{is_restricted}{whether the data.frame should be restricted and only be visible for backend user for (default is TRUE)}
}
\value{
dataset_uid from the created dataset
}
\description{
With this function you can easily update metadata of an existing dataset. You can call the get_metadata() function, edit the value you want and set the new metadata directly from the metadata_df.
}
