% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_dataset_resource.R
\name{update_dataset_resource}
\alias{update_dataset_resource}
\title{Update the dataset resource}
\usage{
update_dataset_resource(filepath, dataset_uid, resource_uid, encoding)
}
\arguments{
\item{filepath}{path to csv file}

\item{dataset_uid}{dataset_uid}

\item{encoding}{encdoing of the csv file}
}
\value{
resource_uid
}
\description{
Update the dataset resource
}
