% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_field_processor.R
\name{update_field_processor}
\alias{update_field_processor}
\title{Update Field Processor}
\usage{
update_field_processor(dataset_uid, body)
}
\arguments{
\item{dataset_uid}{dataset_uid}

\item{body}{body in list and NOT in json format}
}
\description{
Updates or appends a new field configuration
}
