% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_metadata_and_fields.R
\name{update_metadata_and_fields}
\alias{update_metadata_and_fields}
\title{Bestehenden Datensatz bearbeiten (Metadaten und Feldkonfigs)}
\usage{
update_metadata_and_fields(dataset_uid, filepath)
}
\arguments{
\item{dataset_uid}{dataset_uid}

\item{filepath}{Pfad zum Excel Schema}
}
\description{
Bestehenden Datensatz bearbeiten (Metadaten und Feldkonfigs)
}
