% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_resource.R
\name{update_resource}
\alias{update_resource}
\title{update_resource}
\usage{
update_resource(
  dataset_uid,
  resource_uid = NULL,
  filepath,
  encoding,
  separator = NULL
)
}
\arguments{
\item{dataset_uid}{the internal uid of the dataset. The dataset_uid can be taken from the catalog}

\item{resource_uid}{the internal uid of a specific resource (namely a csv file or an attachment)}

\item{filepath}{path to the csv file where the actual data is stored}

\item{encoding}{File encoding (for example utf-8)}

\item{separator}{}
}
\description{
Resource eines Datensatzes ersetzen mit bereits hochgeladenem File
}
