% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_description.R
\name{add_description}
\alias{add_description}
\title{Add a description}
\usage{
add_description(dataset_uid, field, description)
}
\arguments{
\item{dataset_uid}{the internal uid of the dataset. The dataset_uid can be taken from the catalog}

\item{field}{ID of field (variablename in the data.frame)}

\item{description}{description as string}
}
\value{
processor_uid
}
\description{
Add a description
}
