% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_file_to_dataset.R
\name{add_file_to_dataset}
\alias{add_file_to_dataset}
\title{Add CSV File to dataset}
\usage{
add_file_to_dataset(filepath, dataset_uid, encoding, separator = NULL)
}
\arguments{
\item{filepath}{path to the csv file where the actual data is stored}

\item{dataset_uid}{the internal uid of the dataset. The dataset_uid can be taken from the catalog}

\item{encoding}{encdoing of the csv file}

\item{separator}{}
}
\value{
resource_uid
}
\description{
Adds a local CSV file to a dataset on ODS (only possible with automation API)
}
