% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_metadata_from_scheme.R
\name{add_metadata_from_scheme}
\alias{add_metadata_from_scheme}
\title{add_metadata_from_scheme}
\usage{
add_metadata_from_scheme(schema, save_local = TRUE)
}
\arguments{
\item{schema}{path to the metadata xlsx schema}

\item{save_local}{if `TRUE`, the catalog will be saved locally}
}
\value{
dataset_uid
}
\description{
Erstellt neuen Datensatz mit entsprechender Kennung und befüllt diesen mit Metadaten
Funktion kann nur auf genau definiertes Schema angewendet werden.
Alle Metadaten im Schema werden auf das Portal gesladen.
}
