% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_configurations.R
\name{add_timeserie_precision}
\alias{add_timeserie_precision}
\title{Add timeserie precision}
\usage{
add_timeserie_precision(dataset_uid, field, timeserie)
}
\arguments{
\item{dataset_uid}{dataset_uid}

\item{field}{field_id}

\item{timeserie}{timeserie (check opendatasoft for available timeseries)}
}
\value{
processor_uid
}
\description{
Add timeserie precision
}
