% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_configurations.R
\name{add_type}
\alias{add_type}
\title{Change the data type of a field}
\usage{
add_type(dataset_uid, field, type)
}
\arguments{
\item{dataset_uid}{dataset_uid}

\item{field}{field_id}

\item{type}{type (check opendatasoft for available types)}
}
\value{
processor_uid
}
\description{
Change the data type of a field
}
