% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_unit.R
\name{add_unit}
\alias{add_unit}
\title{Add a Unit}
\usage{
add_unit(dataset_uid, field, unit)
}
\arguments{
\item{dataset_uid}{the internal uid of the dataset. The dataset_uid can be taken from the catalog}

\item{field}{ID of field (variablename in the data.frame)}

\item{unit}{unit (check opendatasoft for available units)}
}
\value{
processor_uid
}
\description{
Add a Unit
}
