% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_configurations.R
\name{change_identifier}
\alias{change_identifier}
\title{Change the field ID}
\usage{
change_identifier(dataset_uid, field, to_name)
}
\arguments{
\item{dataset_uid}{dataset_uid}

\item{field}{old field_id}

\item{to_name}{new_field_id}
}
\value{
processor_uid
}
\description{
Change the field ID
}
