% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_dataset.R
\name{delete_dataset}
\alias{delete_dataset}
\title{Datensatz löschen}
\usage{
delete_dataset(dataset_uid, ask = TRUE)
}
\arguments{
\item{dataset_uid}{the internal uid of the dataset. The dataset_uid can be taken from the catalog}

\item{ask}{wenn TRUE wird nach Bestätigung durch USer gefragt.}
}
\description{
Funktion um Datensatz zu löschen
}
