% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_fields.R
\name{edit_fields}
\alias{edit_fields}
\title{Feldkonfigurationen hinzufügen}
\usage{
edit_fields(dataset_uid, schema, original_names = NULL)
}
\arguments{
\item{dataset_uid}{the internal uid of the dataset. The dataset_uid can be taken from the catalog}

\item{schema}{path to the metadata xlsx schema}

\item{original_names}{Namen aus dem `data_info.rds`. default ist `NULL`}
}
\description{
Feldkonfigurationen hinzufügen
}
