% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_fields.R
\name{edit_fields}
\alias{edit_fields}
\title{Feldkonfigurationen hinzufügen}
\usage{
edit_fields(dataset_uid, schema, original_names = NULL)
}
\arguments{
\item{dataset_uid}{the internal uid of the dataset. The dataset_uid can be taken from the catalog}

\item{schema}{path to the metadata xlsx schema}

\item{original_names}{Namen aus dem `data_info.rds`. default ist `NULL`}

\item{filepath}{path to the csv file where the actual data is stored}

\item{save_local}{if `TRUE`, the catalog will be saved locally}

\item{resource_uid}{the internal uid of a specific resource (namely a csv file or an attachment)}

\item{change_uid}{the internal uid of a change}

\item{encoding}{File encoding (for example utf-8)

Attachments}

\item{directory}{Pfad zu Ordner welcher alle Anlagen enthält (Wenn nicht NULL, werden alle Inhalte des Ordners hochgeladen)}

\item{files}{eines oder mehrere Files die als Anlage hochgeladen werden

Metadata}

\item{template}{kann metadata_catalog entnommen werden}

\item{meta_name}{kann metadata_catalog entnommen werden}

\item{meta_value}{neuer Wert}
}
\description{
Feldkonfigurationen hinzufügen
}
