% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_catalog.R
\name{get_catalog}
\alias{get_catalog}
\title{Get detailed Meta Data}
\usage{
get_catalog(dataset_uid = NULL, variables = NULL)
}
\arguments{
\item{dataset_uid}{the internal uid of the dataset. The dataset_uid can be taken from the catalog}

\item{variables}{list of variables that should be downloaded}
}
\value{
dataframe containing the meta data.
}
\description{
Function to retrieve meta data from the backend. This function is for internal use only. An API key must be specified and passed to the function.
}
