% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_catalog.R
\name{get_catalog}
\alias{get_catalog}
\title{Get detailed Meta Data}
\usage{
get_catalog(dataset_uid = NULL)
}
\arguments{
\item{dataset_uid}{dataset_uid}
}
\value{
dataframe containing the meta data.
}
\description{
Function to retrieve meta data from the backend. This function is for internal use only. An API key must be specified and passed to the function.
}
