% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset_info.R
\name{get_dataset_info}
\alias{get_dataset_info}
\title{get_dataset_info}
\usage{
get_dataset_info(save_local = TRUE, dataset_uid = NULL)
}
\arguments{
\item{save_local}{if `TRUE`, the catalog will be saved locally}

\item{dataset_uid}{the internal uid of the dataset. The dataset_uid can be taken from the catalog}
}
\description{
Funktion um alle derzeit aufgeschalteten Metadaten herunterzuladen. Hier kann auch die dataset_uid entnommen werden.
Der Catalog wird automatisch in der Variable metadata_catalog gespeichert
}
