% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{get_metadata}
\usage{
get_metadata(dataset_uid, as_list = FALSE)
}
\arguments{
\item{dataset_uid}{the internal uid of the dataset. The dataset_uid can be taken from the catalog}

\item{as_list}{if `TRUE` the function will return the original JSON as list instead of as data.frame (default is `FALSE`)}
}
\value{
Datensatz mit Metadaten eines gewünschten Datensatzes
}
\description{
Funktion um Metadaten eines bestimmten Datensatzes zu erhalten
}
