% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_configurations.R
\name{make_field_sortable}
\alias{make_field_sortable}
\title{Make field sortable}
\usage{
make_field_sortable(dataset_uid, field)
}
\arguments{
\item{dataset_uid}{dataset_uid}

\item{field}{field_id}
}
\value{
processor_uid
}
\description{
Make field sortable
}
