% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_unique_id.R
\name{mark_unique_id}
\alias{mark_unique_id}
\title{Mark field as unique ID}
\usage{
mark_unique_id(dataset_uid, field)
}
\arguments{
\item{dataset_uid}{the internal uid of the dataset. The dataset_uid can be taken from the catalog}

\item{field}{ID of field (variablename in the data.frame)}
}
\value{
processor_uid
}
\description{
Mark field as unique ID
}
