% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_field.R
\name{rename_field}
\alias{rename_field}
\title{Rename field}
\usage{
rename_field(dataset_uid, field, to_name)
}
\arguments{
\item{dataset_uid}{the internal uid of the dataset. The dataset_uid can be taken from the catalog}

\item{field}{ID of field (variablename in the data.frame)}

\item{to_name}{new name}
}
\value{
processor_uid
}
\description{
Rename field
}
