% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore_changes.R
\name{restore_change}
\alias{restore_change}
\title{restore_change}
\usage{
restore_change(dataset_uid, change_uid)
}
\arguments{
\item{dataset_uid}{Datensatz ID}

\item{change_uid}{Change ID}
}
\description{
Historie wiederherstellen
}
