% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_metadata.R
\name{set_metadata}
\alias{set_metadata}
\title{set_metadata}
\usage{
set_metadata(dataset_uid, template, meta_name, meta_value)
}
\arguments{
\item{dataset_uid}{the internal uid of the dataset. The dataset_uid can be taken from the catalog}

\item{template}{kann metadata_catalog entnommen werden}

\item{meta_name}{kann metadata_catalog entnommen werden}

\item{meta_value}{neuer Wert}
}
\value{
`TRUE` if call was successfull, `NULL` otherwise
}
\description{
Funktion um einen Eintrag in den Metadaten zu ändern
}
