% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpublish_dataset.R
\name{unpublish_dataset}
\alias{unpublish_dataset}
\title{Datensatz Veröffentlihung rückgängig machen}
\usage{
unpublish_dataset(dataset_uid)
}
\arguments{
\item{dataset_uid}{kann metadata_catalog entnommen werden}
}
\description{
Funktion unpublished Datensatz, verändert aber NICHT die Sicherheitseinstellung (Datensatz nur für zugelassene Benutzer sichtbar)
}
