% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_metadata_and_fields.R
\name{update_metadata_and_fields}
\alias{update_metadata_and_fields}
\title{Bestehenden Datensatz bearbeiten (Metadaten und Feldkonfigs)}
\usage{
update_metadata_and_fields(dataset_uid, schema)
}
\arguments{
\item{dataset_uid}{the internal uid of the dataset. The dataset_uid can be taken from the catalog}

\item{schema}{path to the metadata xlsx schema}
}
\description{
Bestehenden Datensatz bearbeiten (Metadaten und Feldkonfigs)
}
