% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_file_to_ods.R
\name{upload_file_to_ods}
\alias{upload_file_to_ods}
\title{Uploads a file to ODS}
\usage{
upload_file_to_ods(dataset_uid, filepath)
}
\arguments{
\item{dataset_uid}{the internal uid of the dataset. The dataset_uid can be taken from the catalog}

\item{filepath}{path to the csv file where the actual data is stored}
}
\value{
results list
}
\description{
Uploads a file to ODS
}
